requirements.txt
streamlit
python-dotenv
requests
langchain
langchain-openai
langchain-community
alpaca-py
pypdf
faiss-cpu
